Fix compiler errors on va_list copies
On amd64, gcc defines va_list as a one-element array of structs, as per
the SYS V amd64 ABI. This requires the use of va_copy(3) instead of
simply assigning the values. Therefore, macros for va_copy and va_end
are provided in fmtdef.h, which use va_copy(3) and va_end(3) if the
former is available.

Signed-off-by: Florian Limberger <flo@purplekraken.com>